Function lexical_core::f32toa_range

#[no_mangle]
pub unsafe extern "C" fn f32toa_range(
    value: f32, 
    first: *mut u8, 
    last: *mut u8
) -> *mut u8

Serializer for a number-to-string conversion using pointer ranges.

Returns a pointer to the 1-past-the-last-byte-written, so that the range [first, last) contains the written bytes. No null-terminator is written.

The data in the range may be uninitialized, these values are never read, only written to.

value - Number to serialize.

first - Pointer to the start of the buffer to write to.

last - Pointer to the one-past-the-end of the buffer to write to.

Panics

Panics if the buffer is not of sufficient size, The caller must provide a range of sufficient size, and neither pointer may be null. In order to ensure the function will not panic, ensure the buffer has at least MAX_*_SIZE elements, using the proper constant for the serialized type from the lexical_core crate root.