Function lexical_core::parse_partial_lossy

pub fn parse_partial_lossy<N: FromLexicalLossy>(
    bytes: &[u8]
) -> Result<(N, usize)>

Lossily parse number from string.

This method parses until an invalid digit is found (or the end of the string), returning the number of processed digits and the parsed value until that point. This parser is lossy, so numerical rounding may occur during parsing.

bytes - Byte slice containing a numeric string.