Function lexical_core::set_exponent_default_char

pub unsafe fn set_exponent_default_char(ch: u8)

Set the default character for the exponent symbol.

Default character for scientific notation, used when the radix < 15.

To change the expected, default character for an exponent, change this value before using lexical.

ch - Character for exponent symbol.

Safety

Do not call this function in threaded-code, as it is not thread-safe. Do not call this function after compiling any formats.

Panics

Panics if the character is in the character set [A-Da-d.+\-].