Function lexical_core::try_atoi128_slice

pub fn try_atoi128_slice(bytes: &[u8]) -> Result<i128>

Checked parser for a string-to-number conversion using Rust slices.

Returns a C-compatible result containing the parsed value, and an error container any errors that occurred during parser.

Numeric overflow takes precedence over the presence of an invalid digit, and therefore may mask an invalid digit error.

bytes - Slice containing a numeric string.