Function lexical_core::try_atoi32_range

#[no_mangle]
pub unsafe extern "C" fn try_atoi32_range(
    first: *const u8, 
    last: *const u8
) -> Result<i32>

Checked parser for a string-to-number conversion using Rust pointer ranges.

Returns a C-compatible result containing the parsed value, and an error container any errors that occurred during parser.

Numeric overflow takes precedence over the presence of an invalid digit, and therefore may mask an invalid digit error.

first - Pointer to the start of the input data.

last - Pointer to the one-past-the-end of the input data.

Panics

Panics if either pointer is null.