Function lexical_core::write_radix

pub fn write_radix<'a, N: ToLexical>(
    n: N,
    radix: u8,
    bytes: &'a mut [u8]
) -> &'a mut [u8]

Write number to string with a custom radix.

Returns a subslice of the input buffer containing the written bytes, starting from the same address in memory as the input slice.

value - Number to serialize.

radix - Radix for number encoding.

bytes - Slice containing a numeric string.

Panics

Panics if the radix is not in the range [2, 36].

Also panics if the buffer may not be large enough to hold the serialized number. In order to ensure the function will not panic, provide a buffer with at least FORMATTED_SIZE elements.