Struct lexical_core::parse_float_options::OptionsBuilder

pub struct OptionsBuilder { /* fields omitted */ }

Builder for Options.

Implementations

impl OptionsBuilder

pub const fn new() -> OptionsBuilder

Create new options builder with default options.

pub const fn get_lossy(&self) -> bool

Get if we disable the use of arbitrary-precision arithmetic.

pub const fn get_exponent(&self) -> u8

Get the character to designate the exponent component of a float.

pub const fn get_decimal_point(&self) -> u8

Get the character to separate the integer from the fraction components.

pub const fn get_nan_string(&self) -> Option<&'static [u8]>

Get the string representation for NaN.

pub const fn get_inf_string(&self) -> Option<&'static [u8]>

Get the short string representation for Infinity.

pub const fn get_infinity_string(&self) -> Option<&'static [u8]>

Get the long string representation for Infinity.

pub const fn lossy(self, lossy: bool) -> OptionsBuilder

Set if we disable the use of arbitrary-precision arithmetic.

pub const fn exponent(self, exponent: u8) -> OptionsBuilder

Set the character to designate the exponent component of a float.

pub const fn decimal_point(self, decimal_point: u8) -> OptionsBuilder

Set the character to separate the integer from the fraction components.

pub const fn nan_string(
self,
nan_string: Option<&'static [u8]>
) -> OptionsBuilder

Set the string representation for NaN.

pub const fn inf_string(
self,
inf_string: Option<&'static [u8]>
) -> OptionsBuilder

Set the short string representation for Infinity.

pub const fn infinity_string(
self,
infinity_string: Option<&'static [u8]>
) -> OptionsBuilder

Set the long string representation for Infinity.

pub const fn nan_str_is_valid(&self) -> bool

Determine if nan_str is valid.

pub const fn inf_str_is_valid(&self) -> bool

Determine if inf_str is valid.

pub const fn infinity_string_is_valid(&self) -> bool

Determine if infinity_string is valid.

pub const fn is_valid(&self) -> bool

Check if the builder state is valid.

pub const unsafe fn build_unchecked(&self) -> Options

Build the Options struct without validation.

Safety

Always safe, just marked as unsafe for API compatibility. The result may be invalid if is_valid is not true.

pub const fn build(&self) -> Result<Options, Error>

Build the Options struct.