Struct lexical_core::WriteFloatOptions

pub struct WriteFloatOptions { /* private fields */ }

Options to customize writing floats.

Examples

use lexical_write_float::Options;

let options = Options::builder()
    .trim_floats(true)
    .nan_string(Some(b"NaN"))
    .inf_string(Some(b"Inf"))
    .build()
    .unwrap();

Implementations

impl Options

pub const fn new() -> Options

Create options with default values.

pub const fn is_valid(&self) -> bool

Check if the options state is valid.

pub const fn max_significant_digits(&self) -> Option<NonZeroUsize>

Get the maximum number of significant digits to write.

pub const fn min_significant_digits(&self) -> Option<NonZeroUsize>

Get the minimum number of significant digits to write.

pub const fn positive_exponent_break(&self) -> Option<NonZeroI32>

Get the maximum exponent prior to using scientific notation.

pub const fn negative_exponent_break(&self) -> Option<NonZeroI32>

Get the minimum exponent prior to using scientific notation.

pub const fn round_mode(&self) -> RoundMode

Get the rounding mode for writing digits with precision control.

pub const fn trim_floats(&self) -> bool

Get if we should trim a trailing ".0" from floats.

pub const fn exponent(&self) -> u8

Get the character to designate the exponent component of a float.

pub const fn decimal_point(&self) -> u8

Get the character to separate the integer from the fraction components.

pub const fn nan_string(&self) -> Option<&'static [u8]>

Get the string representation for NaN.

pub const fn inf_string(&self) -> Option<&'static [u8]>

Get the short string representation for Infinity.

pub unsafe fn set_max_significant_digits(
&mut self,
max_significant_digits: Option<NonZeroUsize>
)

Set the maximum number of significant digits to write. Unsafe, use the builder API for option validation.

Safety

Always safe, just marked as unsafe for API compatibility.

pub unsafe fn set_min_significant_digits(
&mut self,
min_significant_digits: Option<NonZeroUsize>
)

Set the minimum number of significant digits to write. Unsafe, use the builder API for option validation.

Safety

Always safe, just marked as unsafe for API compatibility.

pub unsafe fn set_positive_exponent_break(
&mut self,
positive_exponent_break: Option<NonZeroI32>
)

Set the maximum exponent prior to using scientific notation.

Safety

Always safe, just marked as unsafe for API compatibility.

pub unsafe fn set_negative_exponent_break(
&mut self,
negative_exponent_break: Option<NonZeroI32>
)

Set the minimum exponent prior to using scientific notation.

Safety

Always safe, just marked as unsafe for API compatibility.

pub unsafe fn set_round_mode(&mut self, round_mode: RoundMode)

Set the rounding mode for writing digits with precision control.

Safety

Always safe, just marked as unsafe for API compatibility.

pub unsafe fn set_trim_floats(&mut self, trim_floats: bool)

Set if we should trim a trailing ".0" from floats. Unsafe, use the builder API for option validation.

Safety

Always safe, just marked as unsafe for API compatibility.

pub unsafe fn set_exponent(&mut self, exponent: u8)

Set the character to designate the exponent component of a float.

Safety

Always safe, but may produce invalid output if the exponent is not a valid ASCII character.

pub unsafe fn set_decimal_point(&mut self, decimal_point: u8)

Set the character to separate the integer from the fraction components.

Safety

Always safe, but may produce invalid output if the decimal point is not a valid ASCII character.

pub unsafe fn set_nan_string(&mut self, nan_string: Option<&'static [u8]>)

Set the string representation for NaN. Unsafe, use the builder API for option validation.

Safety

Unsafe if nan_string.len() > MAX_SPECIAL_STRING_LENGTH. This might cause a special string larger than the buffer length to be written, causing a buffer overflow, potentially a severe security vulnerability.

pub unsafe fn set_inf_string(&mut self, inf_string: Option<&'static [u8]>)

Set the short string representation for Infinity Unsafe, use the builder API for option validation.

Safety

Unsafe if nan_string.len() > MAX_SPECIAL_STRING_LENGTH. This might cause a special string larger than the buffer length to be written, causing a buffer overflow, potentially a severe security vulnerability.

pub const fn builder() -> OptionsBuilder

Get WriteFloatOptionsBuilder as a static function.

pub const fn rebuild(&self) -> OptionsBuilder

Create OptionsBuilder using existing values.