Trait lexical_core::ToLexical

pub trait ToLexical: Number {
    fn to_lexical<'a>(self, bytes: &'a mut [u8]) -> &'a mut [u8];
}

Trait for numerical types that can be serialized to bytes.

To determine the number of bytes required to serialize a value to string, check the associated constants from a required trait:

FORMATTED_SIZE

FORMATTED_SIZE_DECIMAL

Required Methods

fn to_lexical<'a>(self, bytes: &'a mut [u8]) -> &'a mut [u8]

Serializer for a number-to-string conversion.

Returns a subslice of the input buffer containing the written bytes, starting from the same address in memory as the input slice.

value - Number to serialize.

bytes - Slice containing a numeric string.

Panics

Panics if the buffer is not of sufficient size. The caller must provide a slice of sufficient size. In order to ensure the function will not panic, ensure the buffer has at least FORMATTED_SIZE_DECIMAL elements.