lexical_core

Module write_float_options

Configuration options for writing floats.

This enables extensive control over how the float is written, from control characters like the decimal point, to the use of exponent notation, and the number of significant digits.

Examples

For example, to customize the writing of numbers for new exponent and decimal point characters, you would use Options to create a custom format and write the integer using the format.

use lexical_write_float::{FormattedSize, Options, ToLexicalWithOptions};
use lexical_write_float::format::STANDARD;

let value = 1.234e45f64;

const CUSTOM: Options = Options::builder()
    // write exponents as "1.2^10" and not "1.2e10"
    .exponent(b'^')
    // use the European decimal point, so "1,2" and not "1.2"
    .decimal_point(b',')
    .build_strict();

const BUFFER_SIZE: usize = CUSTOM.buffer_size_const::<f64, STANDARD>();
let mut buffer = [0u8; BUFFER_SIZE];
let digits = value.to_lexical_with_options::<STANDARD>(&mut buffer, &CUSTOM);
assert_eq!(str::from_utf8(digits), Ok("1,234^45"));

Pre-Defined Formats

These are the pre-defined formats for parsing numbers from various programming, markup, and data languages.

STANDARD: Standard number format.

DECIMAL_COMMA: Numerical format with a decimal comma.

HEX_FLOAT: Numerical format for hexadecimal floats, which use a p exponent.

CARAT_EXPONENT: Numerical format where ^ is used as the exponent notation character.

RUST_LITERAL: Number format for a Rust literal floating-point number.

PYTHON_LITERAL: Number format for a Python literal floating-point number.

CXX_LITERAL: Number format for a C++ literal floating-point number.

C_LITERAL: Number format for a C literal floating-point number.

RUBY_LITERAL: Number format for a Ruby literal floating-point number.

RUBY_STRING: Number format to parse a Ruby float from string.

SWIFT_LITERAL: Number format for a Swift literal floating-point number.

GO_LITERAL: Number format for a Golang literal floating-point number.

HASKELL_LITERAL: Number format for a Haskell literal floating-point number.

HASKELL_STRING: Number format to parse a Haskell float from string.

JAVASCRIPT_LITERAL: Number format for a Javascript literal floating-point number.

JAVASCRIPT_STRING: Number format to parse a Javascript float from string.

PERL_LITERAL: Number format for a Perl literal floating-point number.

PHP_LITERAL: Number format for a PHP literal floating-point number.

JAVA_LITERAL: Number format for a Java literal floating-point number.

JAVA_STRING: Number format to parse a Java float from string.

R_LITERAL: Number format for an R literal floating-point number.

KOTLIN_LITERAL: Number format for a Kotlin literal floating-point number.

KOTLIN_STRING: Number format to parse a Kotlin float from string.

JULIA_LITERAL: Number format for a Julia literal floating-point number.

CSHARP_LITERAL: Number format for a C# literal floating-point number.

CSHARP_STRING: Number format to parse a C# float from string.

KAWA_LITERAL: Number format for a Kawa literal floating-point number.

KAWA_STRING: Number format to parse a Kawa float from string.

GAMBITC_LITERAL: Number format for a Gambit-C literal floating-point number.

GAMBITC_STRING: Number format to parse a Gambit-C float from string.

GUILE_LITERAL: Number format for a Guile literal floating-point number.

GUILE_STRING: Number format to parse a Guile float from string.

CLOJURE_LITERAL: Number format for a Clojure literal floating-point number.

CLOJURE_STRING: Number format to parse a Clojure float from string.

ERLANG_LITERAL: Number format for an Erlang literal floating-point number.

ERLANG_STRING: Number format to parse an Erlang float from string.

ELM_LITERAL: Number format for an Elm literal floating-point number.

ELM_STRING: Number format to parse an Elm float from string.

SCALA_LITERAL: Number format for a Scala literal floating-point number.

SCALA_STRING: Number format to parse a Scala float from string.

ELIXIR_LITERAL: Number format for an Elixir literal floating-point number.

ELIXIR_STRING: Number format to parse an Elixir float from string.

FORTRAN_LITERAL: Number format for a FORTRAN literal floating-point number.

D_LITERAL: Number format for a D literal floating-point number.

COFFEESCRIPT_LITERAL: Number format for a Coffeescript literal floating-point number.

COFFEESCRIPT_STRING: Number format to parse a Coffeescript float from string.

COBOL_LITERAL: Number format for a COBOL literal floating-point number.

COBOL_STRING: Number format to parse a COBOL float from string.

FSHARP_LITERAL: Number format for an F# literal floating-point number.

VB_LITERAL: Number format for a Visual Basic literal floating-point number.

VB_STRING: Number format to parse a Visual Basic float from string.

OCAML_LITERAL: Number format for an OCaml literal floating-point number.

OBJECTIVEC_LITERAL: Number format for an Objective-C literal floating-point number.

OBJECTIVEC_STRING: Number format to parse an Objective-C float from string.

REASONML_LITERAL: Number format for an ReasonML literal floating-point number.

MATLAB_LITERAL: Number format for a MATLAB literal floating-point number.

ZIG_LITERAL: Number format for a Zig literal floating-point number.

SAGE_LITERAL: Number format for a Sage literal floating-point number.

JSON: Number format for a JSON literal floating-point number.

TOML: Number format for a TOML literal floating-point number.

YAML: Number format for a YAML literal floating-point number.

XML: Number format for an XML literal floating-point number.

SQLITE: Number format for a SQLite literal floating-point number.

POSTGRESQL: Number format for a PostgreSQL literal floating-point number.

MYSQL: Number format for a MySQL literal floating-point number.

MONGODB: Number format for a MongoDB literal floating-point number.

Structs

Options

Options to customize writing floats.

OptionsBuilder

Builder for Options.

Enums

RoundMode

Enumeration for how to round floats with precision control.

Constants

CARAT_EXPONENT

Numerical format where ^ is used as the exponent notation character.

CLOJURE_LITERAL

Number format for a Clojure literal floating-point number.

CLOJURE_STRING

Number format to parse a Clojure float from string.

COBOL_LITERAL

Number format for a COBOL literal floating-point number.

COBOL_STRING

Number format to parse a COBOL float from string.

COFFEESCRIPT_LITERAL

Number format for a Coffeescript literal floating-point number.

COFFEESCRIPT_STRING

Number format to parse a Coffeescript float from string.

CSHARP_LITERAL

Number format for a C# literal floating-point number.

CSHARP_STRING

Number format to parse a C# float from string.

CXX_LITERAL

Number format for a C++ literal floating-point number.

C_LITERAL

Number format for a C literal floating-point number.

DECIMAL_COMMA

Numerical format with a decimal comma.

D_LITERAL

Number format for a D literal floating-point number.

ELIXIR_LITERAL

Number format for an Elixir literal floating-point number.

ELIXIR_STRING

Number format to parse an Elixir float from string.

ELM_LITERAL

Number format for an Elm literal floating-point number.

ELM_STRING

Number format to parse an Elm float from string.

ERLANG_LITERAL

Number format for an Erlang literal floating-point number.

ERLANG_STRING

Number format to parse an Erlang float from string.

FORTRAN_LITERAL

Number format for a FORTRAN literal floating-point number.

FSHARP_LITERAL

Number format for an F# literal floating-point number.

GAMBITC_LITERAL

Number format for a Gambit-C literal floating-point number.

GAMBITC_STRING

Number format to parse a Gambit-C float from string.

GO_LITERAL

Number format for a Golang literal floating-point number.

GUILE_LITERAL

Number format for a Guile literal floating-point number.

GUILE_STRING

Number format to parse a Guile float from string.

HASKELL_LITERAL

Number format for a Haskell literal floating-point number.

HASKELL_STRING

Number format to parse a Haskell float from string.

HEX_FLOAT

Numerical format for hexadecimal floats, which use a p exponent.

JAVASCRIPT_LITERAL

Number format for a Javascript literal floating-point number.

JAVASCRIPT_STRING

Number format to parse a Javascript float from string.

JAVA_LITERAL

Number format for a Java literal floating-point number.

JAVA_STRING

Number format to parse a Java float from string.

JSON

Number format for a JSON literal floating-point number.

JULIA_LITERAL

Number format for a Julia literal floating-point number.

KAWA_LITERAL

Number format for a Kawa literal floating-point number.

KAWA_STRING

Number format to parse a Kawa float from string.

KOTLIN_LITERAL

Number format for a Kotlin literal floating-point number.

KOTLIN_STRING

Number format to parse a Kotlin float from string.

MATLAB_LITERAL

Number format for a MATLAB literal floating-point number.

MAX_SPECIAL_STRING_LENGTH

Maximum length for a special string.

MONGODB

Number format for a MongoDB literal floating-point number.

MYSQL

Number format for a MySQL literal floating-point number.

OBJECTIVEC_LITERAL

Number format for an Objective-C literal floating-point number.

OBJECTIVEC_STRING

Number format to parse an Objective-C float from string.

OCAML_LITERAL

Number format for an OCaml literal floating-point number.

PERL_LITERAL

Number format for a Perl literal floating-point number.

PHP_LITERAL

Number format for a PHP literal floating-point number.

POSTGRESQL

Number format for a PostgreSQL literal floating-point number.

PYTHON_LITERAL

Number format for a Python literal floating-point number.

REASONML_LITERAL

Number format for an ReasonML literal floating-point number.

RUBY_LITERAL

Number format for a Ruby literal floating-point number.

RUBY_STRING

Number format to parse a Ruby float from string.

RUST_LITERAL

Number format for a Rust literal floating-point number.

R_LITERAL

Number format for an R literal floating-point number.

SAGE_LITERAL

Number format for a Sage literal floating-point number.

SCALA_LITERAL

Number format for a Scala literal floating-point number.

SCALA_STRING

Number format to parse a Scala float from string.

SQLITE

Number format for a SQLite literal floating-point number.

STANDARD

Standard number format.

SWIFT_LITERAL

Number format for a Swift literal floating-point number.

TOML

Number format for a TOML literal floating-point number.

VB_LITERAL

Number format for a Visual Basic literal floating-point number.

VB_STRING

Number format to parse a Visual Basic float from string.

XML

Number format for an XML literal floating-point number.

YAML

Number format for a YAML literal floating-point number.

ZIG_LITERAL

Number format for a Zig literal floating-point number.