Trait lexical_core::ToLexical

pub trait ToLexical: FormattedSize + Number {
    unsafe fn to_lexical_unchecked<'a>(self, bytes: &'a mut [u8]) -> &'a mut [u8] ;
    fn to_lexical<'a>(self, bytes: &'a mut [u8]) -> &'a mut [u8] ;
}

Trait for numerical types that can be serialized to bytes.

To determine the number of bytes required to serialize a value to string, check the associated constants from a required trait:

FORMATTED_SIZE

FORMATTED_SIZE_DECIMAL

Required Methods

unsafe fn to_lexical_unchecked<'a>(self, bytes: &'a mut [u8]) -> &'a mut [u8]

Serializer for a number-to-string conversion.

Returns a subslice of the input buffer containing the written bytes, starting from the same address in memory as the input slice.

value - Number to serialize.

bytes - Buffer to write number to.

Safety

Safe as long as the caller has provided a buffer of at least FORMATTED_SIZE_DECIMAL elements. If a smaller buffer is provided, a buffer overflow is very likely.

fn to_lexical<'a>(self, bytes: &'a mut [u8]) -> &'a mut [u8]

Serializer for a number-to-string conversion.

Returns a subslice of the input buffer containing the written bytes, starting from the same address in memory as the input slice.

value - Number to serialize.

bytes - Buffer to write number to.

Panics

Panics if the buffer is not of sufficient size. The caller must provide a slice of sufficient size. In order to ensure the function will not panic, ensure the buffer has at least FORMATTED_SIZE_DECIMAL elements.